Попытки сделать принтер

+++ src/main/java/ru/kovalchuk/Impl/command/DeleteCommand.java
@@ -1,7 +1,9 @@
 package ru.kovalchuk.Impl.command;
 
+import lombok.extern.slf4j.Slf4j;
 import ru.kovalchuk.TaskDao;
 
+@Slf4j
 public class DeleteCommand extends BaseCommand {
 
     private int id;
@@ -14,6 +16,14 @@ public class DeleteCommand extends BaseCommand {
 
     @Override
     public void execute() {
-        taskDao.deleteTask(id);
+        if (taskDao.getById(id) != null) {
+            try {
+                taskDao.deleteTask(id);
+            } catch (IndexOutOfBoundsException e) {
+                log.error("Задачи с таким id нет");
+            } catch (NumberFormatException e){
+                log.error("Необходимо ввести id задачи цифрами");
+            }
+        }
     }
 }
